extract test cli
