+++ Cargo.toml
@@ -6,6 +6,7 @@ license = "Apache-2.0"
 homepage = "https://golem.cloud"
 repository = "https://github.com/golemcloud/golem-examples"
 description = "Golem example templates"
+default-run = "golem-examples-cli"
 
 [dependencies]
 clap = { version = "4.4.17", features = ["derive"], optional = true }
@@ -36,4 +37,9 @@ path = "src/lib.rs"
 [[bin]]
 name = "golem-examples-cli"
 path = "src/main.rs"
+required-features = ["cli"]
+
+[[bin]]
+name = "golem-examples-test-cli"
+path = "src/test/main.rs"
 required-features = ["cli"]
