+++ src/cli.rs
@@ -54,14 +54,6 @@ pub enum Command {
         #[arg(short, long, alias = "lang")]
         language: Option<GuestLanguage>,
     },
-
-    // Generates and builds the examples components
-    #[command()]
-    TestExamples {
-        // Filter examples by name, checks if the example name contains the filter string
-        #[arg(short, long)]
-        filter: Option<String>,
-    },
 }
 
 #[derive(Parser, Debug)]
